# job-filter-rest/src/main/resources/openapi/job-filter-api.yml
openapi: 3.0.3
info:
  title: LinkedIn Job Filter API
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /api/users:
    post:
      operationId: createUser
      summary: Create user
      tags: [ Users ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'
        '400':
          description: User creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUserResponse'

  /api/users/{userId}/search-profiles:
    get:
      operationId: getSearchProfiles
      summary: Get search profiles
      tags: [ Search Profiles ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Search profiles returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSearchProfilesResponse'

    post:
      operationId: createSearchProfile
      summary: Create search profile
      tags: [ Search Profiles ]
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSearchProfileRequest'
      responses:
        '201':
          description: Search profile created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSearchProfileResponse'
        '400':
          description: Search profile creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSearchProfileResponse'

  /api/users/{userId}/search-profiles/{searchProfileId}:
    get:
      operationId: getSearchProfile
      summary: Get search profile
      tags: [ Search Profiles ]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/SearchProfileId'
      responses:
        '200':
          description: Search profile returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetSearchProfileResponse'

    put:
      operationId: updateSearchProfile
      summary: Update search profile
      tags: [ Search Profiles ]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/SearchProfileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSearchProfileRequest'
      responses:
        '200':
          description: Search profile updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateSearchProfileResponse'

    delete:
      operationId: deleteSearchProfile
      summary: Delete search profile
      tags: [ Search Profiles ]
      parameters:
        - $ref: '#/components/parameters/UserId'
        - $ref: '#/components/parameters/SearchProfileId'
      responses:
        '200':
          description: Search profile deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmptyResponse'

  /api/job-scans/basic:
    post:
      operationId: submitBasicJobScan
      summary: Submit basic job scan
      tags: [ Job Scans ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BasicJobScanRequest'
      responses:
        '200':
          description: Job scan completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicJobScanResponse'
        '400':
          description: Job scan failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BasicJobScanResponse'

components:
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64

    SearchProfileId:
      name: searchProfileId
      in: path
      required: true
      schema:
        type: integer
        format: int64

  schemas:
    ApiError:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string

    Errors:
      type: array
      items:
        $ref: '#/components/schemas/ApiError'

    EmptyResponse:
      type: object
      required:
        - errors
      properties:
        data:
          nullable: true
        errors:
          $ref: '#/components/schemas/Errors'

    CreateUserResponse:
      type: object
      required:
        - errors
      properties:
        id:
          type: integer
          format: int64
        errors:
          $ref: '#/components/schemas/Errors'

    UpdateSearchProfileResponse:
      type: object
      required:
        - errors
      properties:
        id:
          type: integer
          format: int64
        errors:
          $ref: '#/components/schemas/Errors'

    GetSearchProfileResponse:
      type: object
      required:
        - errors
      properties:
        searchProfile:
          $ref: '#/components/schemas/SearchProfileDTO'
        errors:
          $ref: '#/components/schemas/Errors'

    GetSearchProfilesResponse:
      type: object
      required:
        - errors
      properties:
        searchProfiles:
          type: array
          items:
            $ref: '#/components/schemas/SearchProfileDTO'
        errors:
          $ref: '#/components/schemas/Errors'


    CreateUserRequest:
      type: object
      required:
        - user
      properties:
        user:
          $ref: '#/components/schemas/UserDTO'

    UserDTO:
      type: object
      required:
        - id
        - email
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        displayName:
          type: string

    CreateSearchProfileRequest:
      type: object
      required:
        - searchProfile
      properties:
        searchProfile:
          $ref: '#/components/schemas/SearchProfileDTO'


    CreateSearchProfileResponse:
      type: object
      required:
        - errors
      properties:
        id:
          type: integer
          format: int64
        errors:
          $ref: '#/components/schemas/Errors'

    UpdateSearchProfileRequest:
      type: object
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        isDefault:
          type: boolean
        minScore:
          type: integer
          minimum: 0
          maximum: 100
        idealJobDescription:
          type: string
        configValues:
          type: array
          items:
            $ref: '#/components/schemas/SearchProfileExtendedFieldDTO'

    SearchProfileExtendedFieldDTO:
      type: object
      required:
        - fieldKey
        - value
      properties:
        fieldKey:
          type: string
          example: REQUIRED_SKILL
        value:
          type: string
          example: java
        weight:
          type: number
          format: double
          minimum: 1
          maximum: 10
        isRequired:
          type: boolean

    SearchProfileDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        name:
          type: string
        isDefault:
          type: boolean
        minScore:
          type: integer
        customInstructions:
          type: string
        extendedValues:
          type: array
          items:
            $ref: '#/components/schemas/SearchProfileExtendedFieldDTO'

    BasicJobScanRequest:
      type: object
      required:
        - userId
        - searchProfileId
        - jobs
      properties:
        userId:
          type: integer
          format: int64
        searchProfileId:
          type: integer
          format: int64
        jobs:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ExtractedJobCard'

    ExtractedJobCard:
      type: object
      required:
        - title
        - company
        - location
        - postedDate
        - url
      properties:
        title:
          type: string
        company:
          type: string
        location:
          type: string
        postedDate:
          type: string
        url:
          type: string

    BasicJobScanResponse:
      type: object
      required:
        - errors
      properties:
        data:
          $ref: '#/components/schemas/BasicJobScanData'
        errors:
          $ref: '#/components/schemas/Errors'

    BasicJobScanData:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/JobRating'

    JobRating:
      type: object
      properties:
        jobId:
          type: integer
          format: int64
        title:
          type: string
        company:
          type: string
        location:
          type: string
        postedDate:
          type: string
        url:
          type: string
        ratingType:
          type: string
          enum:
            - BASIC
            - DETAILED
        overallScore:
          type: integer
          minimum: 0
          maximum: 100
        decision:
          type: string
          enum:
            - STRONG_MATCH
            - MAYBE
            - HIDE
        titleMatchScore:
          type: integer
          minimum: 0
          maximum: 100
        careerGoalScore:
          type: integer
          minimum: 0
          maximum: 100
        companyCultureScore:
          type: integer
          minimum: 0
          maximum: 100
        remoteWorkScore:
          type: integer
          minimum: 0
          maximum: 100
        stabilityScore:
          type: integer
          minimum: 0
          maximum: 100
        summary:
          type: string
        reasons:
          type: array
          items:
            type: string
        concerns:
          type: array
          items:
            type: string